requirements.txt
# -----------------------------
# Web framework
# -----------------------------
fastapi==0.110.3
uvicorn[standard]==0.29.0

# -----------------------------
# Data models
# -----------------------------
pydantic==2.7.4

# -----------------------------
# OpenAI
# -----------------------------
openai==1.30.5

# -----------------------------
# LlamaIndex (runtime + controlled indexing)
# -----------------------------
llama-index-core==0.10.52.post1
llama-index-llms-openai==0.1.26
llama-index-embeddings-openai==0.1.9

# -----------------------------
# PDF ingestion (indexing only)
# -----------------------------
pypdf==4.2.0

# -----------------------------
# Utilities / retries
# -----------------------------
requests==2.31.0
tenacity==8.5.0